🛠️feat(file): 优化文件上传功能并重构相关代码，引入 MinioUtils.R 类封装上传文件信息。

+++ utopia-server/common/src/main/java/cn/yshujia/domain/dto/FileDTO.java
@@ -7,7 +7,6 @@ import jakarta.validation.constraints.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -19,32 +18,32 @@ import java.util.List;
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class FileDTO extends PageDTO {
-	
+
 	@Positive(groups = UpdateGroup.class, message = "[资源ID] 不能为空！")
 	private Long id;
-	
+
 	@Positive(groups = InsertGroup.class, message = "[标签ID] 不能为空！")
 	private Long labelId;
-	
+
 	@NotBlank(groups = InsertGroup.class, message = "[资源名] 不能为空！")
 	@Size(groups = {InsertGroup.class, UpdateGroup.class}, max = 32, message = "[资源名] 长度不能超过 32！")
 	private String name;
-	
+
 	private String url;
-	
-	private BigDecimal size;
-	
+
+	private Double size;
+
 	@NotNull(groups = InsertGroup.class, message = "[资源标签] 不能为空！")
 	@Size(groups = {InsertGroup.class, UpdateGroup.class}, max = 10, message = "[资源标签] 不能超过 10 个！")
 	private List<String> tags;
-	
+
 	private Boolean enabled;
-	
+
 	@PositiveOrZero(message = "[审核状态] 必须为正整数！")
 	@Max(value = 2, groups = {InsertGroup.class, UpdateGroup.class}, message = "[审核状态] 值范围 [0, 2]！")
 	@Min(value = 0, groups = {InsertGroup.class, UpdateGroup.class}, message = "[审核状态] 值范围 [0, 2]！")
 	private Integer reviewed;
-	
+
 	private Long createBy;
-	
+
 }
